Update post_refine_phenix.sh
updating:
-adding flag for cif files
-adding removing duplicate hetatom script
-add normalize occ flag to exit
-few buggy code fixes

+++ scripts/post_refine_phenix.sh
@@ -1,22 +1,19 @@
 #!/bin/bash
-
+#___________________________SOURCE__________________________________
 source /home/wankowicz/phenix-installer-1.16-3546-intel-linux-2.6-x86_64-centos6/phenix-1.16-3546/phenix_env.sh
+export PHENIX_OVERWRITE_ALL=true
 
-#pdb=$1
+source /home/wankowicz/anaconda3/etc/profile.d/conda.sh
+conda activate 'qfit2.1'
+
 pdb_name=$1
-F=$2
-SF=$3
 echo $pdb_name
-#set elbow = $5
+
+#__________________________________REMOVE DUPLICATE HET ATOMS__________________________________
+remove_duplicates multiconformer_model2.pdb
 
 
-#bspdb='%s\n' "${pdb//.pdb/}" #`basename $pdb .pdb`
-#echo $bspdb
-#bsmtz='%s\n' "${mtz//.mtz/}" #`basename $mtz .mtz`
-
-#echo $bspdb
-
-# DETERMINE RESOLUTION AND (AN)ISOTROPIC REFINEMENT
+#__________________________________DETERMINE RESOLUTION AND (AN)ISOTROPIC REFINEMENT__________________________________
 resrange= phenix.mtz.dump ${pdb_name}.mtz | grep "Resolution range:"
 
 echo $resrange
@@ -30,62 +27,88 @@ else
   adp='adp.individual.isotropic=all'
 fi
 
-#GET CIF FILE
+#__________________________________GET CIF FILE__________________________________
 phenix.ready_set pdb_file_name=multiconformer_model2.pdb
 
-#CHANGE TO IOBS/SIGOBS
-#phenix.reflection_file_converter
-
-
-# REFINE COORDINATES ONLY
-phenix.refine multiconformer_model2.pdb \
+#__________________________________DETERMINE IF THERE ARE LIGANDS__________________________________
+if [ -f "${pdb_name}.ligands.cif" ]; then
+  phenix.refine multiconformer_model2.pdb.fixed \
               ${pdb_name}.mtz \
-              multiconformer_model2.ligands.cif \
+              ${pdb_name}.ligands.cif \
               strategy=individual_sites \
               output.prefix=${pdb_name} \
               output.serial=2 \
               main.number_of_macro_cycles=5 \
-              #refinement.input.xray_data.labels=$F,$SF\
-              write_maps=false --overwrite
+              write_maps=false \
+              --overwrite
+else
+  phenix.refine multiconformer_model2.pdb.fixed \
+              ${pdb_name}.mtz \
+              strategy=individual_sites \
+              output.prefix=${pdb_name} \
+              output.serial=2 \
+              main.number_of_macro_cycles=5 \
+              write_maps=false \
+              --overwrite
+fi
 
-# REFINE UNTIL OCCUPANCIES CONVERGE
+#__________________________________REFINE UNTIL OCCUPANCIES CONVERGE__________________________________
 zeroes=50
 
-while [ $zeroes > 10 ]
-do
-phenix.refine ${pdb_name}_002.pdb ${pdb_name}.mtz \
-              multiconformer_model2.ligands.cif \
+while [ $zeroes -gt 10 ]; do
+  if [[ -e "${pdb_name}.ligands.cif" ]]; then
+        phenix.refine ${pdb_name}_002.pdb ${pdb_name}.mtz \
+              ${pdb_name}.ligands.cif \
               output.prefix=${pdb_name} \
               output.serial=3 \
               strategy="*individual_sites *individual_adp *occupancies" \
               main.number_of_macro_cycles=5 \
-              #refinement.input.xray_data.labels=$F,$SF\
               write_maps=false --overwrite
-
-zeroes=`normalize_occupancies -occ 0.09 ${pdb_name}_003.pdb`
-   #z=`awk '{ if(ubstr($0,1,6)=="CRYST1"||(substr($0,1,4)=="ATOM"||substr($0,1,6)=="HETATM")&&substr($0,57,4)+0<0.09) print $0}' ${pdb_name}_003.pdb | wc`
-   #zeroes=$z[1]
-
-normalize_occupancies -occ 0.09 ${pdb_name}_003.pdb
-mv ${pdb_name}_003_norm.pdb ${pdb_name}_002.pdb
+  else
+        phenix.refine ${pdb_name}_002.pdb ${pdb_name}.mtz \
+              output.prefix=${pdb_name} \
+              output.serial=3 \
+              strategy="*individual_sites *individual_adp *occupancies" \
+              main.number_of_macro_cycles=5 \
+              write_maps=false --overwrite
+  fi
+  zeroes=`normalize_occupancies -occ 0.09 ${pdb_name}_003.pdb`
+  normalize_occupancies -occ 0.09 ${pdb_name}_003.pdb
+  
+  echo 'Post refinement Zeroes:'
+  echo $zeroes
+  
+  if [ ! -f ${pdb_name}_003_norm.pdb ]; then
+     echo 'normalize occupanies did not work!'
+     exit
+  fi
+  mv ${pdb_name}_003_norm.pdb ${pdb_name}_002.pdb
 done
 
-# ADD HYDROGENS
+#__________________________________ADD HYDROGENS__________________________________
 phenix.reduce ${pdb_name}_002.pdb > ${pdb_name}_004.pdb
 
-# FINAL REFINEMENT
+#__________________________________FINAL REFINEMENT__________________________________
+if [[ -e "${pdb_name}.ligands.cif" ]]; then
 phenix.refine ${pdb_name}_004.pdb ${pdb_name}.mtz \
-              multiconformer_model2.ligands.cif \
-              ${pdb_name}_004.pdb \
+              ${pdb_name}.ligands.cif \
               "$adp" \
               output.prefix=${pdb_name} \
               output.serial=5 \
               strategy="*individual_sites *individual_adp *occupancies" \
               main.number_of_macro_cycles=5 \
-              #refinement.input.xray_data.labels=$F,$SF\
-              multiconformer_model2.ligands.cif \
               write_maps=false \
               --overwrite
+else
+  phenix.refine ${pdb_name}_004.pdb ${pdb_name}.mtz \
+              "$adp" \
+              output.prefix=${pdb_name} \
+              output.serial=5 \
+              strategy="*individual_sites *individual_adp *occupancies" \
+              main.number_of_macro_cycles=5 \
+              write_maps=false \
+              --overwrite
+fi
 
 cp ${pdb_name}_005.pdb ${pdb_name}_qFit.pdb
 cp ${pdb_name}_005.mtz ${pdb_name}_qFit.mtz
